refactor: create <TeamManagementContent/> component
